+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Display error if username in login form contains invalid characters.
 
+### Removed
+
+- Removed "Manage subscriptions" item from profile menu.
+
 ### Fixed
 
 - Added space between posts in search results.